Enpg Conversion Works!
It's done!

+++ Graphics/CIE.cs
@@ -13,9 +13,9 @@ namespace Ndst.Graphics {
 
         // Create a CIE color from RGB5.
         public CIELAB(RGB5 rgb) {
-            double r = rgb.R5 / RGB5.MAX_COMPONENT_VALUE;
-            double g = rgb.G5 /RGB5.MAX_COMPONENT_VALUE;
-            double b = rgb.B5 / RGB5.MAX_COMPONENT_VALUE;
+            double r = rgb.R5 / (double)RGB5.MAX_COMPONENT_VALUE;
+            double g = rgb.G5 / (double)RGB5.MAX_COMPONENT_VALUE;
+            double b = rgb.B5 / (double)RGB5.MAX_COMPONENT_VALUE;
             double x, y, z;
 
             r = (r > 0.04045) ? Math.Pow((r + 0.055) / 1.055, 2.4) : r / 12.92;
@@ -26,9 +26,9 @@ namespace Ndst.Graphics {
             y = (r * 0.2126 + g * 0.7152 + b * 0.0722) / 1.00000;
             z = (r * 0.0193 + g * 0.1192 + b * 0.9505) / 1.08883;
 
-            x = (x > 0.008856) ? Math.Pow(x, 1/3) : (7.787 * x) + 16/116;
-            y = (y > 0.008856) ? Math.Pow(y, 1/3) : (7.787 * y) + 16/116;
-            z = (z > 0.008856) ? Math.Pow(z, 1/3) : (7.787 * z) + 16/116;
+            x = (x > 0.008856) ? Math.Pow(x, (double)1/3) : (7.787 * x) + (double)16/116;
+            y = (y > 0.008856) ? Math.Pow(y, (double)1/3) : (7.787 * y) + (double)16/116;
+            z = (z > 0.008856) ? Math.Pow(z, (double)1/3) : (7.787 * z) + (double)16/116;
 
             L = 116 * y - 16;
             A = 500 * (x - y);
@@ -42,17 +42,17 @@ namespace Ndst.Graphics {
             double z = y - B / 200;
             double r, g, b;
 
-            x = 0.95047 * ((x * x * x > 0.008856) ? x * x * x : (x - 16/116) / 7.787);
-            y = 1.00000 * ((y * y * y > 0.008856) ? y * y * y : (y - 16/116) / 7.787);
-            z = 1.08883 * ((z * z * z > 0.008856) ? z * z * z : (z - 16/116) / 7.787);
+            x = 0.95047 * ((x * x * x > 0.008856) ? x * x * x : (x - (double)16/116) / 7.787);
+            y = 1.00000 * ((y * y * y > 0.008856) ? y * y * y : (y - (double)16/116) / 7.787);
+            z = 1.08883 * ((z * z * z > 0.008856) ? z * z * z : (z - (double)16/116) / 7.787);
 
             r = x *  3.2406 + y * -1.5372 + z * -0.4986;
             g = x * -0.9689 + y *  1.8758 + z *  0.0415;
             b = x *  0.0557 + y * -0.2040 + z *  1.0570;
 
-            r = (r > 0.0031308) ? (1.055 * Math.Pow(r, 1/2.4) - 0.055) : 12.92 * r;
-            g = (g > 0.0031308) ? (1.055 * Math.Pow(g, 1/2.4) - 0.055) : 12.92 * g;
-            b = (b > 0.0031308) ? (1.055 * Math.Pow(b, 1/2.4) - 0.055) : 12.92 * b;
+            r = (r > 0.0031308) ? (1.055 * Math.Pow(r, (double)1/2.4) - 0.055) : 12.92 * r;
+            g = (g > 0.0031308) ? (1.055 * Math.Pow(g, (double)1/2.4) - 0.055) : 12.92 * g;
+            b = (b > 0.0031308) ? (1.055 * Math.Pow(b, (double)1/2.4) - 0.055) : 12.92 * b;
 
             return new RGB5() {
                 R5 = (byte)(Math.Max(0, Math.Min(1, r)) * RGB5.MAX_COMPONENT_VALUE),
